[NV-TF] Fix CVE-2022-21731.
Backport of upstream commit 08d7b00c0a5.

+++ tensorflow/core/framework/common_shape_fns.cc
@@ -1586,7 +1586,7 @@ Status ConcatShapeHelper(InferenceContext* c, int start_value_index,
   const int32 concat_dim = concat_dim_t->scalar<int32>()();
 
   // Minimum required number of dimensions.
-  const int min_rank = concat_dim < 0 ? -concat_dim : concat_dim + 1;
+  const int64 min_rank = concat_dim < 0 ? -concat_dim : concat_dim + 1;
 
   ShapeHandle output_before;
   ShapeHandle output_after;
